custom exception implemented

+++ src/main/java/com/stackroute/muzix/service/TrackServiceImpl.java
@@ -1,5 +1,7 @@
 package com.stackroute.muzix.service;
 
+import com.stackroute.muzix.exception.TrackAlreadyExistsException;
+import com.stackroute.muzix.exception.TrackNotFoundException;
 import com.stackroute.muzix.model.Track;
 import com.stackroute.muzix.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,14 +20,18 @@ public class TrackServiceImpl implements TrackService {
 	}
 	
 	@Override
-	public boolean saveTrack(Track track) {
-			trackRepository.save(track);
+	public boolean saveTrack(Track track) throws TrackAlreadyExistsException {
+		if (trackRepository.existsById(track.getId()))
+			throw new TrackAlreadyExistsException("Track id already exists");
+		trackRepository.save(track);
 			return true;
 	}
 	
 	@Override
-	public boolean deleteTrack(int id) {
-			trackRepository.deleteById(id);
+	public boolean deleteTrack(int id) throws TrackNotFoundException{
+		if (!trackRepository.existsById(id))
+			throw new TrackNotFoundException("Track not found");
+		trackRepository.deleteById(id);
 			return true;
 	}
 	
@@ -36,16 +42,16 @@ public class TrackServiceImpl implements TrackService {
 	}
 	
 	@Override
-	public boolean updateTrack(int id, Track track) {
-			if(trackRepository.existsById(id) == true){
-				trackRepository.save(track);
-				return true;
-			}
-		return false;
+	public boolean updateTrack(int id, Track track) throws TrackNotFoundException{
+		if(!trackRepository.existsById(id)){
+			throw new TrackNotFoundException("Track not found");
+		}
+		trackRepository.save(track);
+		return true;
 	}
 	
 	@Override
-	public List<Track> findTrackByName(String name) {
+	public List<Track> findTrackByName(String name){
 		return trackRepository.trackByName(name);
 	}
 }
